+++ src/main/java/com/kalsym/order/service/model/Store.java
@@ -42,6 +42,7 @@ public class Store implements Serializable {
 
     private String paymentType;
     private Integer invoiceSeqNo;
+    private String invoicePrefix;
     
     public String getNameAbreviation() {
         String abbreviation = "";